Move JS functionaity around
